Move some code from k8s.py to k8s_core.py to make k8s.py slightly more manageable. No change in behavior
